refactor conversion priority to not be global
the global Conversion.nextNumber, used to assign sequential priority
numbers to conversions in the order they are constructed, did not work
well with serialization, necessitating unsafeGetNextPriority and
unsafeSetNextPriority, and when saving and loading the game state we had
to store the current nextNumber alongside the actual Game object.

since having nextNumber as a global variable doesn't make sense, it is
now an attribute of the World, and is serialized as part of the world
instead of requiring special handling.

this means that newly-created conversions start without a priority
number, and sequential numbers are assigned to conversions when tiles
are placed into the world.