refactor world constructor
the world constructor no longer takes width and height as parameters,
since they can be computed from the dimensions of the provided grid.
this also means we don't have to serialize width and height.

the world constructor will now also assign priorities to any conversions
in the provided grid's tiles that were not already assigned. this
doesn't matter right now, since the grid is either from a saved game
or consists entirely of wasteland tiles that have no resource
conversions, but it prevents a bug that would happen if the starting
grid included tiles that have conversions.

additionally, the serialization schema for worlds now constrains the
grid array so that "jagged" arrays are not considered valid.