clean up Conversion.toString and remove isFree
this is relatively minor, but Conversion.toString has been on my list of
code-style improvements because its control flow was unclear.

i've also removed the isFree method. If made
sense when we were treating "free" conversions differently than
conversions that require inputs, but since that is no longer the case
it was only being used in the toString method.